Fix C interop portability and cross-build DLL/.so support

+++ gmpmee.h
@@ -308,7 +308,7 @@ typedef struct
   mpz_t n;             /**< integer to be tested */
   mpz_t n_minus_1;     /**< n minus one */
   mpz_t q;             /**< q is defined by n=q*2^k+1 */
-  unsigned long int k; /**< k is defined by n=q*2^k+1 */
+  unsigned long long k; /**< k is defined by n=q*2^k+1 */
   mpz_t y;             /**< y is temporary space */
 } gmpmee_millerrabin_state[1]; /* Magic references. */
 
@@ -615,6 +615,6 @@ gmpmee_array_clear_dealloc(mpz_t *a, size_t len);
  */
 void
 gmpmee_array_urandomb(mpz_t *rop, size_t len, gmp_randstate_t state,
-		      unsigned long int n);
+          unsigned long long n);
 
 #endif /* GMPMEE_H */
